+++ src/classes/job.ts
@@ -523,6 +523,25 @@ export class Job<
   }
 
   /**
+   * Removes child dependency from parent when child is not yet finished
+   *
+   * @returns True if the relationship existed and if it was removed.
+   */
+  async removeChildDependency(): Promise<boolean> {
+    const childDependencyIsRemoved = await this.scripts.removeChildDependency(
+      this.id,
+      this.parentKey,
+    );
+    if (childDependencyIsRemoved) {
+      this.parent = undefined;
+      this.parentKey = undefined;
+      return true;
+    }
+
+    return false;
+  }
+
+  /**
    * Clears job's logs
    *
    * @param keepLogs - the amount of log entries to preserve
@@ -704,7 +723,12 @@ export class Job<
 
     const code = results[results.length - 1][1] as number;
     if (code < 0) {
-      throw this.scripts.finishedErrors(code, this.id, command, 'active');
+      throw this.scripts.finishedErrors({
+        code,
+        jobId: this.id,
+        command,
+        state: 'active',
+      });
     }
 
     if (finishedOn && typeof finishedOn === 'number') {